# Replace `sync.Mutex` with `sync.RWMutex` for CachedEnforcer 