`sync.Map` appears to fit better here as the current implementation serializes any writes to the map even though they are typically writing separate keys (and typically only ever a given key, once). Using `sync.Map` reduces contention on this mutex.

From https://golang.org/pkg/sync/#Map
> The Map type is optimized for two common use cases: (1) when the entry for a given key is only ever written once but read many times, as in caches that only grow, or (2) when multiple goroutines read, write, and overwrite entries for disjoint sets of keys. In these two cases, use of a Map may significantly reduce lock contention compared to a Go map paired with a separate Mutex or RWMutex.

I think this use case fits the first well.

Running the added benchmark locally, I saw a roughly 40% improvement.

Before:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel
2019/04/25 13:35:27 Start automatically load policy     
2019/04/25 13:35:27 Stop automatically load policy
2019/04/25 13:35:27 Start automatically load policy
2019/04/25 13:35:27 Model:
2019/04/25 13:35:27 p.p: sub, obj, act        
2019/04/25 13:35:27 e.e: some(where (p_eft == allow))
2019/04/25 13:35:27 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act     
2019/04/25 13:35:27 r.r: sub, obj, act
2019/04/25 13:35:27 Policy:                   
2019/04/25 13:35:27 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 13:35:27 Request: alice, data1, read ---> true
2019/04/25 13:35:27 Request: alice, data1, write ---> false                                                                                                                                                        
2019/04/25 13:35:27 Request: alice, data2, read ---> false
2019/04/25 13:35:27 Request: alice, data2, write ---> false
2019/04/25 13:35:27 Request: bob, data1, read ---> false
2019/04/25 13:35:27 Request: bob, data1, write ---> false
2019/04/25 13:35:27 Request: bob, data2, read ---> false
2019/04/25 13:35:27 Request: bob, data2, write ---> true
2019/04/25 13:35:27 Stop automatically load policy   
goos: linux                                                                
goarch: amd64              
pkg: github.com/casbin/casbin                                               
BenchmarkCachedRBACModelMediumParallel-4         5000000               359 ns/op
PASS                                                       
ok      github.com/casbin/casbin        5.137s    
```

After:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel
2019/04/25 13:35:11 Start automatically load policy
2019/04/25 13:35:11 Stop automatically load policy                              
2019/04/25 13:35:11 Start automatically load policy
2019/04/25 13:35:11 Model:
2019/04/25 13:35:11 r.r: sub, obj, act
2019/04/25 13:35:11 p.p: sub, obj, act
2019/04/25 13:35:11 e.e: some(where (p_eft == allow))
2019/04/25 13:35:11 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act
2019/04/25 13:35:11 Policy:
2019/04/25 13:35:11 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 13:35:11 Request: alice, data1, read ---> true
2019/04/25 13:35:11 Request: alice, data1, write ---> false
2019/04/25 13:35:11 Request: alice, data2, read ---> false
2019/04/25 13:35:11 Request: alice, data2, write ---> false
2019/04/25 13:35:11 Request: bob, data1, read ---> false
2019/04/25 13:35:11 Request: bob, data1, write ---> false
2019/04/25 13:35:11 Request: bob, data2, read ---> false
2019/04/25 13:35:11 Request: bob, data2, write ---> true
2019/04/25 13:35:11 Stop automatically load policy
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin
BenchmarkCachedRBACModelMediumParallel-4        10000000               220 ns/op
PASS
ok      github.com/casbin/casbin        5.434s
```

I only added a parallel benchmark of BenchmarkCachedRBACModelMedium for now, but
we could opt to instead replace existing benchmarks with parallel benchmarks to show how the enforcer behaves when accessed from multiple goroutines.