@vtolstov 2019-04-25
Can you try test also, if sync.Mutex replaces with sync.RWMutex? as i read from other projects, sync.Map not recommended to use and using own locking with RWMutex gave more performance