@jszwedko 2019-04-25
@vtolstov good thought, it looks like the locking on read contention is the main culprit.

With this diff:

```diff
diff --git a/enforcer_cached.go b/enforcer_cached.go
index fffc834..14cea5e 100644
--- a/enforcer_cached.go
+++ b/enforcer_cached.go
@@ -23,7 +23,7 @@ type CachedEnforcer struct {
        *Enforcer
        m           map[string]bool
        enableCache bool
-       locker      *sync.Mutex
+       locker      *sync.RWMutex
 }

 // NewCachedEnforcer creates a cached enforcer via file or DB.
@@ -32,7 +32,7 @@ func NewCachedEnforcer(params ...interface{}) *CachedEnforcer {
        e.Enforcer = NewEnforcer(params...)
        e.enableCache = true
        e.m = make(map[string]bool)
-       e.locker = new(sync.Mutex)
+       e.locker = new(sync.RWMutex)
        return e
 }

@@ -57,13 +57,16 @@ func (e *CachedEnforcer) Enforce(rvals ...interface{}) bool {
                }
        }

-       e.locker.Lock()
-       defer e.locker.Unlock()
+       e.locker.RLock()
        if _, ok := e.m[key]; ok {
+               e.locker.RUnlock()
                return e.m[key]
        } else {
+               e.locker.RUnlock()
                res := e.Enforcer.Enforce(rvals...)
+               e.locker.Lock()
                e.m[key] = res
+               e.locker.Unlock()
                return res
        }
 }
```

I see:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel                                                                                                        
2019/04/25 19:20:49 Start automatically load policy
2019/04/25 19:20:49 Stop automatically load policy
2019/04/25 19:20:49 Start automatically load policy
2019/04/25 19:20:49 Model:
2019/04/25 19:20:49 r.r: sub, obj, act
2019/04/25 19:20:49 p.p: sub, obj, act
2019/04/25 19:20:49 e.e: some(where (p_eft == allow))
2019/04/25 19:20:49 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act
2019/04/25 19:20:49 Policy:
2019/04/25 19:20:49 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 19:20:49 Request: alice, data1, read ---> true
2019/04/25 19:20:49 Request: alice, data1, write ---> false
2019/04/25 19:20:49 Request: alice, data2, read ---> false
2019/04/25 19:20:49 Request: alice, data2, write ---> false
2019/04/25 19:20:49 Request: bob, data1, read ---> false
2019/04/25 19:20:49 Request: bob, data1, write ---> false
2019/04/25 19:20:49 Request: bob, data2, read ---> false
2019/04/25 19:20:49 Request: bob, data2, write ---> true
2019/04/25 19:20:49 Stop automatically load policy
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin
BenchmarkCachedRBACModelMediumParallel-4        10000000               226 ns/op
PASS
ok      github.com/casbin/casbin        5.593s
```

Roughly the same as using `sync.Map`. I haven't profiled memory usage to see if there are any differences.

Do you have a preference between `sync.RWMutex` or `sync.Map` given this?