@jszwedko 2019-04-25
Benchmarks with memory usage:

Using `sync.Mutex`:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel -benchmem                                                                                                
2019/04/25 19:27:27 Start automatically load policy
2019/04/25 19:27:27 Stop automatically load policy
2019/04/25 19:27:27 Start automatically load policy
2019/04/25 19:27:27 Model:
2019/04/25 19:27:27 r.r: sub, obj, act
2019/04/25 19:27:27 p.p: sub, obj, act
2019/04/25 19:27:27 e.e: some(where (p_eft == allow))
2019/04/25 19:27:27 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act
2019/04/25 19:27:27 Policy:
2019/04/25 19:27:27 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 19:27:27 Request: alice, data1, read ---> true
2019/04/25 19:27:27 Request: alice, data1, write ---> false
2019/04/25 19:27:27 Request: alice, data2, read ---> false
2019/04/25 19:27:27 Request: alice, data2, write ---> false
2019/04/25 19:27:27 Request: bob, data1, read ---> false
2019/04/25 19:27:27 Request: bob, data1, write ---> false
2019/04/25 19:27:27 Request: bob, data2, read ---> false
2019/04/25 19:27:27 Request: bob, data2, write ---> true
2019/04/25 19:27:28 Stop automatically load policy
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin
BenchmarkCachedRBACModelMediumParallel-4         5000000               375 ns/op             128 B/op          4 allocs/op                                                                                         
PASS
ok      github.com/casbin/casbin        5.450s
```

Using `sync.RWMutex`:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel -benchmem
2019/04/25 19:26:25 Start automatically load policy
2019/04/25 19:26:25 Stop automatically load policy
2019/04/25 19:26:25 Start automatically load policy
2019/04/25 19:26:25 Model:
2019/04/25 19:26:25 p.p: sub, obj, act
2019/04/25 19:26:25 e.e: some(where (p_eft == allow))
2019/04/25 19:26:25 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act
2019/04/25 19:26:25 r.r: sub, obj, act
2019/04/25 19:26:25 Policy:
2019/04/25 19:26:25 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 19:26:25 Request: alice, data1, read ---> true
2019/04/25 19:26:25 Request: alice, data1, write ---> false
2019/04/25 19:26:25 Request: alice, data2, read ---> false
2019/04/25 19:26:25 Request: alice, data2, write ---> false
2019/04/25 19:26:25 Request: bob, data1, read ---> false
2019/04/25 19:26:25 Request: bob, data1, write ---> false
2019/04/25 19:26:25 Request: bob, data2, read ---> false
2019/04/25 19:26:25 Request: bob, data2, write ---> true
2019/04/25 19:26:25 Stop automatically load policy
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin
BenchmarkCachedRBACModelMediumParallel-4         5000000               234 ns/op             128 B/op          4 allocs/op
PASS
ok      github.com/casbin/casbin        4.656s
```

Using `sync.Map`:

```
go test -cpu 4 -bench BenchmarkCachedRBACModelMediumParallel -benchmem                                                                                                
2019/04/25 19:28:08 Start automatically load policy
2019/04/25 19:28:08 Stop automatically load policy
2019/04/25 19:28:08 Start automatically load policy
2019/04/25 19:28:08 Model:
2019/04/25 19:28:08 r.r: sub, obj, act
2019/04/25 19:28:08 p.p: sub, obj, act
2019/04/25 19:28:08 e.e: some(where (p_eft == allow))
2019/04/25 19:28:08 m.m: r_sub == p_sub && r_obj == p_obj && r_act == p_act
2019/04/25 19:28:08 Policy:
2019/04/25 19:28:08 p: sub, obj, act: [[alice data1 read] [bob data2 write]]
2019/04/25 19:28:08 Request: alice, data1, read ---> true
2019/04/25 19:28:08 Request: alice, data1, write ---> false
2019/04/25 19:28:08 Request: alice, data2, read ---> false
2019/04/25 19:28:08 Request: alice, data2, write ---> false
2019/04/25 19:28:08 Request: bob, data1, read ---> false
2019/04/25 19:28:08 Request: bob, data1, write ---> false
2019/04/25 19:28:08 Request: bob, data2, read ---> false
2019/04/25 19:28:08 Request: bob, data2, write ---> true
2019/04/25 19:28:09 Stop automatically load policy
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin
BenchmarkCachedRBACModelMediumParallel-4        10000000               248 ns/op             128 B/op          4 allocs/op                                                                                         
PASS
ok      github.com/casbin/casbin        5.959s
```