@jszwedko 2019-04-26
@vtolstov :+1: I force-pushed a commit switching to `sync.RWMutex`. I'm inclined to remove the benchmark I added and just replace the other tests with parallel execution. What do you think?